#642 Split LeaderUtility to RulerUtility
